[PORT FROM MAIN] CBS : Update the "Delete All" display
BZ: 65376

In the CellBroadcastReceiver application, when the user
presses for the first time the menu touch, with one or
more SMS_CB received already, the "Delete All" is not
displayed.

This patch updates the menu display of the
CellBroadcastReceiver application. With this patch,
the "delete all" appears correctly in the settings
menu, if one or several SMS are present in the application.

Change-Id: Ida6c0a546bd3a1b52a6bb5df03ecc866d10b81e8
Orig-Change-Id: I5c655937986ee1a4542ab77dc1fba40104b25bcc
Category: device enablement
Domain: Telephony-Modem-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Delaude Emmanuel <emmanuelx.delaude@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>