[PORT FROM MAIN] CBS : Update the "Delete All" display
BZ: 65376

In the CellBroadcastReceiver application, when the user
presses for the first time the menu touch, with one or
more SMS_CB received already, the "Delete All" is not
displayed.

This patch updates the menu display of the
CellBroadcastReceiver application. With this patch,
the "delete all" appears correctly in the settings
menu, if one or several SMS are present in the application.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I5c655937986ee1a4542ab77dc1fba40104b25bcc
Orig-Change-Id: I51c2ce915965250c7015b6f922b4a719e3c16fd3
Signed-off-by: Delaude Emmanuel <emmanuelx.delaude@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/123902
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>