[PORT FROM MAIN] CBS : Update the "Delete All" display.
BZ: 65376

In the CellBroadcastReceiver application, when the user
presses for the first time the menu touch, with one or
more SMS_CB received already, the "Delete All" is not
displayed.

This patch updates the menu display of the
CellBroadcastReceiver application. With this patch,
the "delete all" appears correctly in the settings
menu, if one or several SMS are present in the application.

Change-Id: I51c2ce915965250c7015b6f922b4a719e3c16fd3
Orig-Change-Id: I9a45f3808f47ac2d118ddff403a99c7dc08e320b
Signed-off-by: Delaude Emmanuel <emmanuelx.delaude@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/79027
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>