cellbroadcast: review implementation of cellbroadcast settings
BZ: 108491

The "Alert sound duration" pop-up of the cellbroadcast
settings disappears after phone rotation.

This issue is due to a bad implementation of the settings
with the PreferenceFragment class. Using this class is not
necessary here and we can have only one class extending
the PreferenceActivity one.

This patch reworks the CellBroadcastSettings class in this
way to fix the issue.

Change-Id: Ifdd844877daeff55f9c4af14c53656a80221d49b
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/111599
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>